Fix b-vlan-tag-vlan-id translation

+++ v3po/v3po2vpp/src/main/java/io/fd/hc2vpp/v3po/interfaces/pbb/PbbRewriteCustomizer.java
@@ -112,8 +112,8 @@ public class PbbRewriteCustomizer extends FutureJVppCustomizer
         return checkNotNull(data.getSourceAddress(), "Destination address cannot be null").getValue();
     }
 
-    private byte verifiedBVlanId(final PbbRewrite data) {
-        return (byte) checkNotNull(data.getBVlanTagVlanId(), "BVlan id cannot be null").shortValue();
+    private short verifiedBVlanId(final PbbRewrite data) {
+        return checkNotNull(data.getBVlanTagVlanId(), "BVlan id cannot be null").shortValue();
     }
 
     private int verifiedISid(final PbbRewrite data) {
